Refactories necessarios, ajustes na estrutura e padronização de constantes.
