+++ .idea/.name
@@ -1,1 +1,1 @@
-Meus Contatos
+android2-lista-contato
