+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/MeusContatos.iml" filepath="$PROJECT_DIR$/MeusContatos.iml" />
+      <module fileurl="file://$PROJECT_DIR$/android2-lista-contato.iml" filepath="$PROJECT_DIR$/android2-lista-contato.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
