+++ app/src/main/java/br/com/meuscontatos/principal/activity/MainActivity.java
@@ -1,10 +1,11 @@
-package br.com.meuscontatos.principal;
+package br.com.meuscontatos.principal.activity;
 
 import android.os.Bundle;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.ActionBarActivity;
 import android.support.v7.widget.Toolbar;
 
+import br.com.meuscontatos.principal.R;
 import br.com.meuscontatos.principal.fragment.ListaContatosFragment;
 import br.com.meuscontatos.principal.fragment.TabContatosPrincipalFragment;
