+++ build.gradle
@@ -5,7 +5,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.1.2'
+        classpath 'com.android.tools.build:gradle:2.1.3'
         classpath 'io.realm:realm-gradle-plugin:1.2.0'
         classpath 'com.getkeepsafe.dexcount:dexcount-gradle-plugin:0.5.2'
     }
